Additional updates for running src tools in fim_run.sh - moved src .py files from /tools to /src - deleted params_calibrated.env (no longer needed) - removed old bathy_rc_adjust.py from /src - deleted dev versions of bathy_src_adjust scripts - added doc strings and cleaned up --help text
