--- config/params_calibrated.env
@@ -1,42 +1,0 @@
-#!/bin/bash
-
-#### hydroconditioning parameters ####
-export negative_burn_value=1000
-export agree_DEM_buffer=70
-export wbd_buffer=5000
-export thalweg_lateral_elev_threshold=3
-
-#### geospatial parameters ####
-export max_split_distance_meters=1500
-export ms_buffer_dist=7000
-export lakes_buffer_dist_meters=20
-
-#### rating curve parameters ####
-export manning_n="/foss_fim/config/mannings_calibrated.json"
-export stage_min_meters=0
-export stage_interval_meters=0.3048
-export stage_max_meters=25
-export slope_min=0.001
-export min_catchment_area=0.25
-export min_stream_length=0.5
-
-#### bathy SRC estimation parameters ####
-export bankfull_input_table="/bathymetry/BANKFULL_CONUS.txt" # Option 1: Wieczorek (2018) bankfull geometry (Bieger 2015 regression)
-#export bankfull_input_table="/bathymetry/nwm_route_link_geom_BED.csv" # Option 2: NWM Route Link bankfull geometry (Blackburn-Lynch regression)
-export bathy_src_modification=True # Toggle to turn BARC on/off (True=on; False=off)
-export surf_area_thalweg_ratio_flag=10 # Flag: Surface area ratio value to identify possible thalweg notch "jump" (SA x+1 / SA x)
-export thalweg_stg_search_max_limit=3 # Threshold: Stage value limit below which to look for the surface area ratio flag (only flag thalweg notch below this threshold)
-export bathy_xs_area_chg_flag=1 # Flag: Cross section area limit to cap the amount of bathy XS area added to the SRC. Limits the bathy_calc_xs_area/ BANKFULL_XSEC_AREA to the specified threshold
-export bankful_xs_area_ratio_flag=10 # Flag: Identify bogus BARC adjusted values where the regression bankfull XS Area/SRC bankfull area is > threshold (topwidth crosswalk issues or bad bankfull regression data points??)
-export thalweg_hyd_radius_flag=10 # Flag: Idenitify possible erroneous BARC-adjusted hydraulic radius values. BARC discharge values greater than the specified threshold and within the thal_stg_limit are set to 0
-export ignore_streamorders=10 # Ignore BARC calculations for streamorders >= this value (10 is Mississippi R)
-
-#### computational parameters ####
-export ncores_gw=1 # mpi number of cores for gagewatershed
-export ncores_fd=1 # mpi number of cores for flow directions
-export default_max_jobs=1 # default number of max concurrent jobs to run
-export memfree=0G # min free memory required to start a new job or keep youngest job alive
-
-#### logging parameters ####
-export startDiv="\n##########################################################################\n"
-export stopDiv="\n##########################################################################"
